Ignore content types without change lists.
This eliminates the limit_choices_to requirement in the content_type field. If
the admin doesn't provide a list of this type of content, then it will be
ignored.

Honestly, I'm not familiar enough with the django.contrib.admin internals to
know why this happens. But, it's common and bypassing the failures provides our
feature for all user defined models.

Fixes #1